+++ sacred/observers/mongo.py
@@ -20,6 +20,9 @@ from sacred.serializer import flatten
 from sacred.utils import ObserverError
 
 
+DEFAULT_MONGO_PRIORITY = 30
+
+
 def force_valid_bson_key(key):
     key = str(key)
     if key.startswith('$'):
@@ -54,7 +57,7 @@ class MongoObserver(RunObserver):
 
     @staticmethod
     def create(url='localhost', db_name='sacred', collection='runs',
-               overwrite=None, **kwargs):
+               overwrite=None, priority=DEFAULT_MONGO_PRIORITY, **kwargs):
         client = pymongo.MongoClient(url, **kwargs)
         database = client[db_name]
         if collection in MongoObserver.COLLECTION_NAME_BLACKLIST:
@@ -62,13 +65,16 @@ class MongoObserver(RunObserver):
                            'Please use a different one.'.format(collection))
         runs_collection = database[collection]
         fs = gridfs.GridFS(database)
-        return MongoObserver(runs_collection, fs, overwrite=overwrite)
+        return MongoObserver(runs_collection, fs, overwrite=overwrite,
+                             priority=priority)
 
-    def __init__(self, runs_collection, fs, overwrite=None):
+    def __init__(self, runs_collection, fs, overwrite=None,
+                 priority=DEFAULT_MONGO_PRIORITY):
         self.runs = runs_collection
         self.fs = fs
         self.overwrite = overwrite
         self.run_entry = None
+        self.priority = priority
 
     def queued_event(self, ex_info, command, queue_time, config, meta_info,
                      _id):
@@ -252,43 +258,55 @@ class MongoDbOption(CommandLineOption):
 
     arg = 'DB'
     arg_description = "Database specification. Can be " \
-                      "[host:port:]db_name[.collection]"
+                      "[host:port:]db_name[.collection][!priority]"
 
-    DB_NAME_PATTERN = r"[_A-Za-z][0-9A-Za-z!#%&'()+\-;=@\[\]^_{}.]{0,63}"
+    DB_NAME_PATTERN = r"[_A-Za-z][0-9A-Za-z#%&'()+\-;=@\[\]^_{}.]{0,63}"
     HOSTNAME_PATTERN = \
         r"(?=.{1,255}$)"\
         r"[0-9A-Za-z](?:(?:[0-9A-Za-z]|-){0,61}[0-9A-Za-z])?"\
         r"(?:\.[0-9A-Za-z](?:(?:[0-9A-Za-z]|-){0,61}[0-9A-Za-z])?)*"\
         r"\.?"
     URL_PATTERN = "(?:" + HOSTNAME_PATTERN + ")" + ":" + "(?:[0-9]{1,5})"
-
-    DB_NAME = re.compile("^" + DB_NAME_PATTERN + "$")
-    URL = re.compile("^" + URL_PATTERN + "$")
+    PRIORITY_PATTERN = "(?P<priority>!-?\d+)?"
+    DB_NAME = re.compile("^" + DB_NAME_PATTERN + PRIORITY_PATTERN + "$")
+    URL = re.compile("^" + URL_PATTERN + PRIORITY_PATTERN + "$")
     URL_DB_NAME = re.compile("^(?P<url>" + URL_PATTERN + ")" + ":" +
-                             "(?P<db_name>" + DB_NAME_PATTERN + ")$")
+                             "(?P<db_name>" + DB_NAME_PATTERN + ")" +
+                             PRIORITY_PATTERN + "$")
 
     @classmethod
     def apply(cls, args, run):
-        url, db_name, collection = cls.parse_mongo_db_arg(args)
+        url, db_name, collection, priority = cls.parse_mongo_db_arg(args)
         if collection:
             mongo = MongoObserver.create(db_name=db_name, url=url,
-                                         collection=collection)
+                                         collection=collection,
+                                         priority=priority)
         else:
-            mongo = MongoObserver.create(db_name=db_name, url=url)
+            mongo = MongoObserver.create(db_name=db_name, url=url,
+                                         priority=priority)
 
         run.observers.append(mongo)
 
     @classmethod
     def parse_mongo_db_arg(cls, mongo_db):
+        def get_priority(pattern):
+            prio_str = pattern.match(mongo_db).group('priority')
+            if prio_str is None:
+                return DEFAULT_MONGO_PRIORITY
+            else:
+                return int(prio_str[1:])
+
         if cls.DB_NAME.match(mongo_db):
             db_name, _, collection = mongo_db.partition('.')
-            return 'localhost:27017', db_name, collection
+            return ('localhost:27017', db_name, collection,
+                    get_priority(cls.DB_NAME))
         elif cls.URL.match(mongo_db):
-            return mongo_db, 'sacred', ''
+            return mongo_db, 'sacred', '', get_priority(cls.URL)
         elif cls.URL_DB_NAME.match(mongo_db):
             match = cls.URL_DB_NAME.match(mongo_db)
             db_name, _, collection = match.group('db_name').partition('.')
-            return match.group('url'), db_name, collection
+            return (match.group('url'), db_name, collection,
+                    get_priority(cls.URL_DB_NAME))
         else:
             raise ValueError('mongo_db argument must have the form "db_name" '
                              'or "host:port[:db_name]" but was {}'