+++ HISTORY.rst
@@ -27,6 +27,7 @@ Release History
 * Feature: added sacred.SETTINGS as a place to configure some of the behaviour
 * Feature: ConfigScopes now extract docstrings and line comments and display
            them when calling ``print_config``
+* Feature: observers are now run in order of priority (settable)
 * API Change: all times are now in UTC
 * API Change: significantly changed the mongoDB layout
 * API Change: MongoObserver and FileStorageObserver now use consecutive